pyproject.toml
[project]
name = "mlflow-practice"
version = "0.1.0"
description = "MLflow 실험 프로젝트"
requires-python = ">=3.10"
dependencies = [
    "mlflow>=3.0.0rc2",
    "psycopg2-binary",
    "pandas",
    "scikit-learn",
    "matplotlib",
    "seaborn",
    "boto3>=1.38.27",
    "python-dotenv>=1.1.0",
    "pip>=25.1.1",
]

[tool.uv]
prerelease = "allow"
